updates Stripe checkout page to receive the order details and calculated order Total from a fetch request to the database using the order id passed in the url. passes order info fetch request to Stripe's Api as well for pricing in payment intent
